% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Receive the raw earthquake dataset and clean it to make it usable for visualisation}
\usage{
eq_clean_data(df_raw_NOAA)
}
\arguments{
\item{df_raw_NOAA}{The raw earthquake dataset}
}
\value{
Returns a clean dataset with an added DATE column
}
\description{
Receive the raw earthquake dataset and clean it to make it usable for visualisation
}
\examples{
\dontrun{
 df_cleaneddata <- eq_clean_data(readr::read_delim("earthquakes.tsv.gz",delim = "\\t"))
}
}
