% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label function}
\usage{
eq_create_label(df_data)
}
\arguments{
\item{df_data}{The data set containing the earthquake information}
}
\value{
A character vector containing the annotation text in html format
}
\description{
This function receives the earthquake data set and builds an annotation_text in html format
that can be displayed as a pop-up on the map. The function checks to see if the LOCATION_NAME,
EQ_PRIMARY and DEATHS for each earthquake contains a valid value and will add the lines if
they are valid. Otherwise, it will simply skip the specific field.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
 eq_clean_data() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}
}
