% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean the LOCATION_NAME column of the earthquake dataset}
\usage{
eq_location_clean(df_data)
}
\arguments{
\item{df_data}{The earthquake dataframe as sent by the eq_clean_data function}
}
\value{
This function returns the earthquake dataframe with a cleaned LOCATION_NAME column in title case
}
\description{
Clean the LOCATION_NAME column of the earthquake dataset
}
\examples{
\dontrun{
 df_cleaneddata <- eq_locationclean(readr::read_delim("earthquakes.tsv.gz",delim = "\\t"))
}
}
