% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map function}
\usage{
eq_map(df_data, annot_col = "DATE")
}
\arguments{
\item{df_data}{The data set containing the earthquake information}

\item{annot_col}{The column that will be used for the pop-up. Default set to DATE}
}
\value{
An interactive leaflet map displaying the earthquakes contained in the dataset
}
\description{
This function receives a data set of earthquake information and plots it
on an interactive map using leaflet. It will use the annot_col parameter
to set the value of the pop-up based column of the data set that matches annot_col.
}
\details{
The earthquakes are mapped using the Latitude and Longitude for each and the size of the circle
reflects the magnitude of the earthquake.
}
\examples{
\dontrun {
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
 eq_map(annot_col = "DATE")
}
}
