% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeLine proto}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
GeomTimeline
}
\description{
This geom is responsible for drawing a line per country and plotting the earthquakes on it.
The magnitude of the earthquake is used for the size of the circles, while the deaths are used for colour.
}
\keyword{datasets}
