% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline function}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetics that the geom accepts}

\item{stat}{The statistical transformation function that must be used}
}
\value{
This function returns a plot where earthquakes are plotted per country and where the size
is the magnitude and the colour is the number of deaths
}
\description{
geom_timeline function
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz",delim = "\\t") \%>\%
eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
 ggplot() +
 geom_timeline(aes(x = DATE, y = COUNTRY,size = EQ_PRIMARY, colour = DEATHS))
}
}
