% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{geom_timeline_label function}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, xmin = NULL, xmax = NULL, n_max = 5, fill = NA)
}
\arguments{
\item{mapping}{The aesthetics that the geom accepts}

\item{stat}{The statistical transformation function that must be used}

\item{n_max}{The maximum number of labels to display}
}
\value{
This function returns the labels that are plotted on the timeline
}
\description{
geom_timeline_label function
}
\examples{
\dontrun {
readr::read_delim("earthquakes.tsv.gz",delim = "\\t") \%>\%
eq_clean_data() \%>\%
 dplyr::filter(COUNTRY == "USA" & lubridate::year(DATE) >= 2000) \%>\%
 ggplot() +
 geom_timeline(aes(x = DATE, y = COUNTRY,size = EQ_PRIMARY, colour = DEATHS)) +
 geom_timeline_label(aes(x = DATE, y = COUNTRY, label = LOCATION_NAME, size = EQ_PRIMARY), n_max = 5) +
 ggtitle("Earthquake Timeline") +
 theme_timeline +
 labs(size = "Richter Scale value:", colour = "# of Deaths:")
}
}
